+++ core/apps.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
 from django.apps import AppConfig
+from creators.apps import CreatorsConfig
 
 
 class CoreConfig(AppConfig):
