--- core/models2/Article.py
@@ -1,7 +1,0 @@
-from django.db import models
-
-class Article(models.Model):
-
-    title = models.CharField(max_length=20)
-    author = models.CharField(max_length=20)
-    content = models.TextField()
