+++ core/templates/sections/designpanel.html
