+++ core/templates/sections/events.html
