+++ core/urls.py
@@ -22,14 +22,11 @@ urlpatterns = [
     url(r'^$', views.index, name='index'),
     url(r'^admin/', admin.site.urls),
     url(r'^auth/', include('social_django.urls', namespace='social')),
-    url(r'^creators/$', views.creators, name='creators'),
-    url(r'^creators_new/$', views.creators_new, name='creators_new'),
-    url(r'^creators_main/$', views.creators_main, name='creators_main'),
-    url(r'^creators_signin/$', views.creators_signin, name='creators_signin'),
     url(r'^events/$', views.events, name='events'),
     url(r'^designpanel/$', views.designpanel, name='designpanel'),
     url(r'^joinus/$', views.joinus, name='joinus'),
     url(r'^login/$', views.login, name='login'),
     url(r'^raw/$', views.rawexpo, name='rawexpo'),
     url(r'^teams/$', views.teams, name='teams'),
+    url(r'^creators/', include('creators.urls', namespace='creators')),
 ]
