+++ core/views.py
@@ -11,26 +11,6 @@ def index(request):
     return render(request, 'sections/index.html', context)
 
 
-def creators(request):
-    context = {}
-    return render(request, 'sections/creatorshome.html', context)
-
-
-def creators_new(request):
-    context = {}
-    return render(request, 'sections/creatorsnewproject.html', context)
-
-
-def creators_signin(request):
-    context = {}
-    return render(request, 'sections/creatorssignin.html', context)
-
-
-def creators_main(request):
-    context = {}
-    return render(request, 'sections/creatorsmain.html', context)
-
-
 def events(request):
     query = Event.objects.order_by('date').all().reverse()
