+++ creators/models/profile.py
@@ -10,8 +10,7 @@ class Profile(models.Model):
     college = models.CharField(max_length=100) #some sort of enum thingy?
     grad_date = models.DateTimeField(blank=True, null=True)
     link = models.CharField(max_length=200)
-
-
-def __str__(self):
-        return self.name
+    
+    def __str__(self):
+            return self.name
