+++ creators/models/tag.py
@@ -1,6 +1,7 @@
 from django.db import models
 
-class Tag(models.Model)
+
+class Tag(models.Model):
     name = models.CharField(max_length=20)
 
     def __str__(self):
