--- creators/static/.gitignore
@@ -1,5 +1,0 @@
-.sass-cache/
-css/
-*.map
-imgs/events
-imgs/sections/events
